third_party/blink/web_tests/virtual/selective-in-order-script/wpt_internal/in-order-script-scheduling/selective-in-order/async-script.sub_default-expected.txt

This is a testharness.js-based test.
[FAIL] InOrder Script Execution Order with allowlist
  Uncaught Error: assert_array_equals: Execution order expected property 1 to be "Sync1" but got "Sync2" (expected array ["Inline1", "Sync1", "Sync2", "EndOfBody"] got ["Inline1", "Sync2", "EndOfBody", "Sync1"])
Harness: the test ran to completion.
